refactor: Component를 core 폴더로 이동하고 initialState를 props안으로 이동
