+++ src/components/UserSection.js
@@ -1,5 +1,5 @@
 import { API_HEADER_X_USERNAME } from "../constants.js";
-import Component from "./Component.js";
+import Component from "../core/Component.js";
 
 export default class UserSection extends Component {
   render() {
